apps/api/requirements.txt
# ─── Web Framework ───────────────────────────────────────────────────────────
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
httpx==0.26.0
aiohttp==3.9.3
aiofiles==23.2.1

# ─── Database ────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.4
psycopg2-binary==2.9.9

# ─── Caching / Queue ─────────────────────────────────────────────────────────
redis==5.0.1
celery==5.3.6
flower==2.0.1
kombu==5.3.4

# ─── Auth ────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.0

# ─── AI / LLM ────────────────────────────────────────────────────────────────
anthropic==0.40.0
openai==1.12.0
google-generativeai==0.4.0
tiktoken==0.6.0

# ─── Document Processing ─────────────────────────────────────────────────────
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
lxml==5.1.0
markdown==3.5.2
html2text==2020.1.16
openpyxl==3.1.2
python-pptx==0.6.23

# ─── Web Crawling ────────────────────────────────────────────────────────────
playwright==1.41.2
scrapy==2.11.0
crawlee==0.0.5

# ─── Embeddings / NLP ────────────────────────────────────────────────────────
sentence-transformers==2.3.1
numpy==1.26.3
scikit-learn==1.4.0
nltk==3.8.1
spacy==3.7.2

# ─── Storage ─────────────────────────────────────────────────────────────────
boto3==1.34.34
python-magic==0.4.27

# ─── Graph ───────────────────────────────────────────────────────────────────
networkx==3.2.1

# ─── Monitoring / Observability ──────────────────────────────────────────────
structlog==24.1.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.40.0

# ─── Utilities ───────────────────────────────────────────────────────────────
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
tenacity==8.2.3
rich==13.7.0
click==8.1.7
croniter==2.0.1
pytz==2024.1
python-dateutil==2.8.2
orjson==3.9.13
ujson==5.9.0
